refactor: rename parityswap-prop exchange to parity-prop
