+++ pkg/liquidity-source/parityprop/abi.go
@@ -1,4 +1,4 @@
-package parityswapprop
+package parityprop
 
 import (
 	"bytes"
