+++ pkg/liquidity-source/parityprop/embed.go
@@ -1,4 +1,4 @@
-package parityswapprop
+package parityprop
 
 import _ "embed"
