+++ pkg/liquidity-source/parityprop/math.go
@@ -1,4 +1,4 @@
-package parityswapprop
+package parityprop
 
 import "github.com/holiman/uint256"
