+++ pkg/liquidity-source/parityprop/math_test.go
@@ -1,4 +1,4 @@
-package parityswapprop
+package parityprop
 
 import (
 	"testing"
