+++ pkg/liquidity-source/parityprop/pool_list_updater_test.go
@@ -1,4 +1,4 @@
-package parityswapprop
+package parityprop
 
 import (
 	"context"
