+++ pkg/liquidity-source/parityprop/pool_simulator_test.go
@@ -1,4 +1,4 @@
-package parityswapprop
+package parityprop
 
 import (
 	"math/big"
