+++ pkg/liquidity-source/parityprop/pool_tracker.go
@@ -1,4 +1,4 @@
-package parityswapprop
+package parityprop
 
 import (
 	"context"
