+++ pkg/liquidity-source/parityprop/type.go
@@ -1,4 +1,4 @@
-package parityswapprop
+package parityprop
 
 import "github.com/holiman/uint256"
